refactor(car): point legacy mutation adapters at the moved query keys
